+++ packages/react-core/src/constants/Integrations.ts
@@ -1,0 +1,10 @@
+// integrations
+export const INTEGRATIONS_OPTION = 'integrations';
+export const INTEGRATIONS_OPENAI_OPTION = 'openai';
+export const INTEGRATIONS_OPENAI_SECRETKEY_OPTION = 'secretKey';
+
+export const DEFAULT_INTEGRATIONS_OPTION = {
+  [INTEGRATIONS_OPENAI_OPTION]: {
+    [INTEGRATIONS_OPENAI_SECRETKEY_OPTION]: null,
+  },
+};