+++ packages/react-ui/src/index.ts
@@ -8,7 +8,11 @@ import AccountDeleteButton, {AccountDeleteButtonProps} from './components/Accoun
 import NavigationSettingsIconButton, {NavigationSettingsIconButtonProps, NavigationSettingsItem} from './components/NavigationSettingsIconButton';
 import NavigationToolbar, {NavigationToolbarProps, NavigationToolbarSkeleton} from './components/NavigationToolbar';
 import NavigationToolbarMobile, {NavigationToolbarMobileProps, NavigationToolbarMobileSkeleton} from './components/NavigationToolbarMobile';
-import NavigationMenuIconButton, {NavigationMenuContent, NavigationMenuIconButtonProps} from './components/NavigationMenuIconButton';
+import NavigationMenuIconButton, {
+  NavigationMenuContent,
+  NavigationMenuIconButtonProps,
+  NavigationMenuHeader
+} from './components/NavigationMenuIconButton';
 import BottomNavigation, {BottomNavigationProps} from './components/BottomNavigation';
 import BroadcastMessages, {BroadcastMessagesProps, BroadcastMessagesSkeleton} from './components/BroadcastMessages';
 import Category, {CategoryProps, CategorySkeleton} from './components/Category';
@@ -111,7 +115,8 @@ import {
   SCUserProfileFields,
   SCUserProfileSettings,
   SCUserSocialAssociations,
-  VirtualScrollerItemProps
+  VirtualScrollerItemProps,
+  PlatformWidgetActionType
 } from './types';
 import Widget, {WidgetProps} from './components/Widget';
 import ConfirmDialog from './shared/ConfirmDialog/ConfirmDialog';
@@ -166,6 +171,8 @@ import UserSubscribedGroupsWidget, {
   UserSubscribedGroupsWidgetSkeleton
 } from './components/UserSubscribedGroupsWidget';
 import DefaultDrawerContent, {DefaultDrawerContentProps} from './components/NavigationMenuIconButton/DefaultDrawerContent';
+import DefaultHeaderContent, {DefaultHeaderContentProps} from './components/NavigationMenuIconButton/DefaultHeaderContent';
+
 /**
  * Constants
  */
@@ -188,6 +195,10 @@ import {getUnseenNotification, getUnseenNotificationCounter} from './utils/feed'
 import * as MessageUploaderUtils from './utils/thumbnailCoverter';
 import {getRelativeTime} from './utils/formatRelativeTime';
 import FeedObjectMediaPreview, {FeedObjectMediaPreviewProps} from './components/FeedObjectMediaPreview';
+/**
+ * Import Assets
+ */
+import LogoSelfCommunity from './assets/logo';
 /**
  * List all exports
  */
@@ -219,6 +230,7 @@ export {
   NavigationToolbarSkeleton,
   NavigationMenuIconButton,
   NavigationMenuContent,
+  NavigationMenuHeader,
   NavigationMenuIconButtonProps,
   BottomNavigation,
   BottomNavigationProps,
@@ -362,6 +374,7 @@ export {
   UserSocialAssociationProps,
   SCUserSocialAssociations,
   VirtualScrollerItemProps,
+  PlatformWidgetActionType,
   CustomAdv,
   CustomAdvProps,
   CustomAdvSkeleton,
@@ -450,6 +463,8 @@ export {
   UserSubscribedGroupsWidgetSkeleton,
   DefaultDrawerContent,
   DefaultDrawerContentProps,
+  DefaultHeaderContent,
+  DefaultHeaderContentProps,
   /* SC UI SHARED */
   HiddenPlaceholder,
   UrlTextField,
@@ -526,5 +541,8 @@ export {
   BaseDialog,
   BaseDialogProps,
   GroupSettingsIconButton,
-  GroupSettingsIconButtonProps
+  GroupSettingsIconButtonProps,
+
+  /* Assets */
+  LogoSelfCommunity
 };