+++ packages/react-templates/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.51-alpha.5](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.3.51-alpha.4...@selfcommunity/react-templates@0.3.51-alpha.5) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.3.51-alpha.4](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.3.51-alpha.3...@selfcommunity/react-templates@0.3.51-alpha.4) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.3.51-alpha.3](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.3.51-alpha.2...@selfcommunity/react-templates@0.3.51-alpha.3) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.3.51-alpha.2](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.3.51-alpha.1...@selfcommunity/react-templates@0.3.51-alpha.2) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.3.51-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.3.51-alpha.0...@selfcommunity/react-templates@0.3.51-alpha.1) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
+## [0.3.51-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.3.50...@selfcommunity/react-templates@0.3.51-alpha.0) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/react-templates
+
+
+
+
+
 ## [0.3.50](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-templates@0.3.50-alpha.1...@selfcommunity/react-templates@0.3.50) (2024-07-10)
 
 **Note:** Version bump only for package @selfcommunity/react-templates