+++ packages/react-theme-default/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.51-alpha.5](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-theme-default@0.1.51-alpha.4...@selfcommunity/react-theme-default@0.1.51-alpha.5) (2024-07-25)
+
+
+### Bug Fixes
+
+* DefaultDrawerContent, DefaultHeaderContent ([1032c1f](https://github.com/selfcommunity/community-js/commit/1032c1fb0c79b3b8cc696bfb835ba40fe262aab0))
+
+
+
+
+
+## [0.1.51-alpha.4](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-theme-default@0.1.51-alpha.3...@selfcommunity/react-theme-default@0.1.51-alpha.4) (2024-07-25)
+
+**Note:** Version bump only for package @selfcommunity/react-theme-default
+
+
+
+
+
+## [0.1.51-alpha.3](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-theme-default@0.1.51-alpha.2...@selfcommunity/react-theme-default@0.1.51-alpha.3) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/react-theme-default
+
+
+
+
+
+## [0.1.51-alpha.2](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-theme-default@0.1.51-alpha.1...@selfcommunity/react-theme-default@0.1.51-alpha.2) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/react-theme-default
+
+
+
+
+
+## [0.1.51-alpha.1](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-theme-default@0.1.51-alpha.0...@selfcommunity/react-theme-default@0.1.51-alpha.1) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/react-theme-default
+
+
+
+
+
+## [0.1.51-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-theme-default@0.1.50...@selfcommunity/react-theme-default@0.1.51-alpha.0) (2024-07-19)
+
+
+### Bug Fixes
+
+* init platform widget ([679bcb7](https://github.com/selfcommunity/community-js/commit/679bcb729998c11723db34b2bac1e0d15d4e7d4d))
+* platform widget with tutorial steps ([959dcdb](https://github.com/selfcommunity/community-js/commit/959dcdb444118c665038e916a6b448f9eb1bc0ee))
+
+
+
+
+
 ## [0.1.50](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-theme-default@0.1.50-alpha.1...@selfcommunity/react-theme-default@0.1.50) (2024-07-10)
 
 **Note:** Version bump only for package @selfcommunity/react-theme-default