+++ packages/types/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.51-alpha.0](https://github.com/selfcommunity/community-js/compare/@selfcommunity/types@0.5.50...@selfcommunity/types@0.5.51-alpha.0) (2024-07-19)
+
+**Note:** Version bump only for package @selfcommunity/types
+
+
+
+
+
 ## [0.5.50](https://github.com/selfcommunity/community-js/compare/@selfcommunity/types@0.5.50-alpha.1...@selfcommunity/types@0.5.50) (2024-07-10)
 
 **Note:** Version bump only for package @selfcommunity/types